requirements.txt
# Core AI and ML libraries
openai>=1.0.0
anthropic>=0.7.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Web framework and API
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.0
aiohttp>=3.8.0

# Data processing and storage
chromadb>=0.4.0
langchain>=0.0.200

# Image and multimodal support
pillow>=10.0.0
opencv-python>=4.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0
colorama>=0.4.6
